+++ server/controllers/notes.js
@@ -1,30 +1,50 @@
 const Notes = require("../models/Notes");
 
 async function getAllNotes(req, res) {
-    const allNotes = await Notes.find({}).sort({createdAt: -1});
-    res.json({allNotes});
+    try {
+        const allNotes = await Notes.find({creatorId: req.userId}).sort({createdAt: -1});
+        res.json({allNotes});
+    } catch (err) {
+        res.json({error: "Unexpected error"});
+    }
 }
 
 async function createNote(req, res) {
-    const {title, note} = req.body;
-    const newNote = await Notes.create({title, note});
-    res.json({newNote});
+    try {
+        const {title, note} = req.body;
+        const newNote = await Notes.create({title, note, creatorId: req.userId});
+        res.json({newNote});
+    } catch (err) {
+        res.json({error: "Unexpected error"});
+    }
 }
 
 async function getNote(req, res) {
-    const note = await Notes.findOne({_id: req.params.id});
-    res.json({note});
+    try {
+        const note = await Notes.findOne({_id: req.params.id, creatorId: req.userId});
+        res.json({note});
+    } catch (err) {
+        res.json({error: "Unexpected error"});
+    }
 }
 
 async function updateNote(req, res) {
-    const {title, note} = req.body;
-    const updatedNote = await Notes.findOneAndUpdate({_id: req.params.id}, {title, note}, {new: true});
-    res.json({updatedNote});
+    try {
+        const {title, note} = req.body;
+        const updatedNote = await Notes.findOneAndUpdate({_id: req.params.id, creatorId: req.userId}, {title, note}, {new: true});
+        res.json({updatedNote});
+    } catch (err) {
+        res.json({error: "Unexpected error"});
+    }
 }
 
 async function deleteNote(req, res) {
-    const deletedNote = await Notes.findOneAndDelete({_id: req.params.id});
-    res.json({deletedNote});
+    try {
+        const deletedNote = await Notes.findOneAndDelete({_id: req.params.id, creatorId: req.userId});
+        res.json({deletedNote});
+    } catch (err) {
+        res.json({error: "Unexpected error"});
+    }
 }
 
 module.exports = {
